requirements.txt
streamlit==1.55.0
groq==1.1.1
yfinance==1.2.0
pandas==3.0.1
numpy==2.4.3
plotly==6.6.0
requests==2.32.5
python-dotenv==1.2.2
pytest==9.0.2
pytest-asyncio==1.3.0
# Database
chromadb==0.6.3
sentence-transformers==3.4.1
google-genai==1.3.0
fredapi==0.5.2
transformers==4.49.0
torch==2.6.0